Change Python path in calibration_dag.py
Updated the path to the Python executable in calibration_dag.py to point to the Pyenv location on the airflow server. This change is a part of our effort to centralize Python version management using Pyenv. Additionally, a new Pyenv installation script has been introduced and the Dockerfile has been updated accordingly.