Remove WORKDIR command from Dockerfile
The WORKDIR command has been removed from the Dockerfile. The removal of the command does not affect the operation of the Docker container, and the following commands will still be executed in the correct locations. Moreover, the Python path in the calibration_dag.py has been updated to centralize Python version management using Pyenv.