Update pyenv installation paths in Dockerfile
The installation paths for pyenv have been updated to '/home/airflow/.bashrc' in the Dockerfile. This change adjusts where the pyenv initializations are echoed to during Docker image building. Ensuring pyenv is correctly set up for the 'airflow' user allows for a smoother user experience for those utilizing the Docker image.