Update python path in calibration_dag.py
The python path for the "external_python" task in the calibration_dag.py file has been corrected. Previously, the path was starting with a double slash, which has been adjusted to a single slash. This change ensures accurate task execution by referencing the correct python version installed with pyenv in the 'airflow' user's environment.