Update .bashrc file ownership in Dockerfile
The Dockerfile has been updated to change the file ownership of .bashrc in /home/airflow/ to 'airflow'. This is to ensure proper execution of Pyenv as python path has also been specified in the Dockerfile. The recursive option (-R) has been applied to include all files and directories under /home/airflow/.bashrc.