Update user switching and pip install commands in Dockerfile
The changes shift the user switching commands to an earlier point in the Dockerfile to ensure subsequent commands are run under this user. In addition, the pip install commands have been adjusted to use the '-q' option to simplify the output, thus facilitating the identification of any issues during the Docker build.