Remove --no-cache option from Docker build command
The --no-cache command line option was removed from the Docker build command in the Makefile. This change will utilize Docker's caching system which speeds up the build process by reusing the intermediate layers from previous builds, as long as the Dockerfile hasn't changed. It can be more performance-efficient but any issues related to stale data might not be caught.